[AMDGPU] Improve scheduling model for VOP3b instructions
VOP3b instructions like v_addc_u32 write vcc (an sgpr) as well as a vgpr
result. The way this was modelled made the write to vcc take an extra
micro-op, which made the whole instruction take twice as long to issue,
which is inaccurate.

Fix this by introducing a new write class that doesn't consume any
resources.

Differential Revision: https://reviews.llvm.org/D75910