AMDGPU/SelectionDAG: Don't combine uniform multiplies to MUL_[UI]24
TODO:
- Judge remaining test changes

Prefer to keep uniform (non-divergent) multiplies on the scalar ALU when
possible. This significantly improves some game cases.

Since isDivergent is only an approximation of whether a value is in SGPRs,
it can potentially regress some situations where a uniform value ends up
in a VGPR. These should be rare in real code, although the test changes
do contain a number of examples.

GlobalISel appears to handle more scenarios in the desirable way,
although it can also be thrown off and fails to select the 24-bit
multiplies in some cases.

Alternative solution considered and rejected was to allow selecting
MUL_[UI]24 to S_MUL_I32. I've reject this because the definition of
those SD operations works is don't-care on the most significant 8 bits,
and this fact is used in some combines via SimplifyDemandedBits.

Change-Id: I728f7a295b48c3686de4837174a8b7356b77bbc2